+++ boxsdk/__init__.py
@@ -3,5 +3,7 @@
 from __future__ import unicode_literals
 
 from .auth import JWTAuth, OAuth2
-from .client import Client, DeveloperTokenClient
+from .client import *  # pylint:disable=wildcard-import,redefined-builtin
 from .object import *  # pylint:disable=wildcard-import,redefined-builtin
+from .config import Version
+__version__ = Version.VERSION