+++ boxsdk/config.py
@@ -9,3 +9,9 @@ class API(object):
     UPLOAD_URL = 'https://upload.box.com/api/2.0'
     OAUTH2_API_URL = 'https://api.box.com/oauth2'  # <https://developers.box.com/docs/#oauth-2>
     OAUTH2_AUTHORIZE_URL = 'https://app.box.com/api/oauth2/authorize'  # <https://developers.box.com/docs/#oauth-2-authorize>
+
+
+class Version(object):
+    """Configuration object containing the package version and user agent string."""
+    VERSION = '1.4.3'
+    USER_AGENT_STRING = 'box-python-sdk-{0}'.format(VERSION)